+++ android/core/domain/src/main/java/com/ohdodok/catchytape/core/domain/usecase/login/AutomaticallyLoginUseCase.kt
@@ -1,4 +1,4 @@
-package com.ohdodok.catchytape.core.domain.usecase
+package com.ohdodok.catchytape.core.domain.usecase.login
 
 import com.ohdodok.catchytape.core.domain.repository.AuthRepository
 import com.ohdodok.catchytape.core.domain.repository.UserTokenRepository
@@ -8,6 +8,7 @@ class AutomaticallyLoginUseCase @Inject constructor(
     private val authRepository: AuthRepository,
     private val userTokenRepository: UserTokenRepository
 ) {
+
     suspend operator fun invoke(): Boolean {
         val accessToken = userTokenRepository.getAccessToken()
         if (accessToken.isBlank()) return false
