+++ android/core/domain/src/main/java/com/ohdodok/catchytape/core/domain/usecase/login/LoginUseCase.kt
@@ -1,4 +1,4 @@
-package com.ohdodok.catchytape.core.domain.usecase
+package com.ohdodok.catchytape.core.domain.usecase.login
 
 import com.ohdodok.catchytape.core.domain.repository.AuthRepository
 import com.ohdodok.catchytape.core.domain.repository.UserTokenRepository
@@ -6,14 +6,13 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
 
-class SignUpUseCase @Inject constructor(
+class LoginUseCase @Inject constructor(
     private val authRepository: AuthRepository,
     private val userTokenRepository: UserTokenRepository
 ) {
 
-    operator fun invoke(googleToken: String, nickname: String): Flow<Unit> =
-        authRepository.signUpWithGoogle(googleToken, nickname).map {
+    operator fun invoke(googleToken: String): Flow<Unit> =
+        authRepository.loginWithGoogle(googleToken).map {
             userTokenRepository.saveAccessToken(it)
         }
-
 }
