+++ android/feature/login/src/main/java/com/ohdodok/catchytape/feature/login/NicknameViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.ohdodok.catchytape.core.domain.usecase.signup.NicknameValidationResult
 import com.ohdodok.catchytape.core.domain.usecase.signup.ValidateNicknameUseCase
-import com.ohdodok.catchytape.core.domain.usecase.SignUpUseCase
+import com.ohdodok.catchytape.core.domain.usecase.signup.SignUpUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
